Complete session 1 and 2 challenges
