+++ session1/3-challenge/1_arithmetic.rb
@@ -8,7 +8,5 @@
 # arithmetic1(10.5)   # => 32.5
 # arithmetic1(-6)     # => -50
 
-def arithmetic1(n)
-  n * 5 - 20
-end
+
