+++ session1/3-challenge/2_arithmetic.rb
@@ -8,6 +8,4 @@
 # arithmetic2(19, 10)  # => 5.0
 # arithmetic2(-6, -7)  # => -3.5
 
-def arithmetic2(a, b)
-  a < b ? a.to_f / 2 : b.to_f / 2
-end
+
