+++ session1/3-challenge/3_simple_logic.rb
@@ -7,6 +7,4 @@
 # ten_twenty(5) # => 20
 # ten_twenty(6) # => 10
 
-def ten_twenty(n)
-  n.even? ? 10 : 20
-end
+
