+++ session1/3-challenge/5_string.rb
@@ -4,6 +4,3 @@
 # add_more_ruby("The sad dad said sad stuff.")  # => "The happy dad said happy stuff."
 # add_more_ruby("Sad times are ahead!")         # => "Happy times are ahead!"
 
-def add_more_ruby(string)
-  string.gsub!(/sad/, "happy"); string.gsub!(/Sad/, "Happy"); string
-end
