+++ session1/3-challenge/6_string.rb
@@ -5,16 +5,3 @@
 # If you have no idea where to begin, remember to check out the cheatsheets for string and logic/control
 #
 
-def odds_and_evens(string, return_odds)
-  mystring = ""
-  if return_odds
-    string.size.times do |i|
-      i.even? ? next : mystring << string[i]
-    end
-  else
-    string.size.times do |i|
-      i.odd? ? next : mystring << string[i]
-    end
-  end
-  mystring
-end
