+++ session1/5-solved/3.rb
@@ -1,7 +1,3 @@
 def ten_twenty(n)
-  if n % 2 == 0
-    10
-  else
-    20
-  end
+  n.even? ? 10 : 20
 end
