+++ session1/5-solved/5.rb
@@ -1,3 +1,3 @@
 def add_more_ruby(string)
-  string.gsub('sad', 'happy').gsub('Sad', 'Happy')
+  string.gsub!(/sad/, "happy"); string.gsub!(/Sad/, "Happy"); string
 end
