+++ session2/3-challenge/10_classes.rb
@@ -15,16 +15,4 @@
 # josh.age      # => 30
 #
 
-class Person
-  attr_accessor :name
-  attr_accessor :age
 
-  def initialize(name, age)
-    @name = name
-    @age = age
-  end
-
-  def birthday
-    @age += 1
-  end
-end
