+++ session2/3-challenge/1_input_output.rb
@@ -7,9 +7,3 @@
 # Standard input will be like "9 2\n" and will expect you to print
 # "11\n7\n18\n" to standard output.
 
-def sum_difference_product
-  x, y = gets.split.map {|n| n.to_i }
-  puts x + y
-  puts x - y
-  puts x * y
-end
