+++ session2/3-challenge/2_input_output_control.rb
@@ -19,15 +19,7 @@
 # and when you think it is correct, you can test it with  $ rake 2:2
 
 
-def hi_hi_goodbye
-  prompt = "Enter a number"
-  while (number = gets) && (number !~ /bye/)
-    number.to_i.times do (print "hi ")
-    end
-    prompt
-  end
-  puts "goodbye"
-end
+
