+++ session2/3-challenge/3_array.rb
@@ -5,12 +5,4 @@
 # "abcdefg".every_other_char  # => "aceg"
 # "".every_other_char         # => ""
 
-class String
-  def every_other_char
-    mystring = ""
-    each_char.with_index do |char, index|
-      index.even? ? mystring << char : nil
-    end
-    mystring
-  end
-end
+
