+++ session2/3-challenge/4_array.rb
@@ -8,7 +8,4 @@
 
 # This time you will have to define the method, it's called: get_squares
 
-def get_squares(numbers)
-  numbers.select {|num| numbers.include? num ** 2 }.sort
-end
