+++ session2/3-challenge/5_array.rb
@@ -14,8 +14,3 @@
 #
 # mod_three [0,1,2,3,4,5,6,7] # => [1, 2, 1, 2, 1]
 
-def mod_three(numbers)
-  myary = []
-  numbers.map {|num| myary << num % 3 unless num % 3 == 0 }
-  myary
-end
