+++ session2/3-challenge/6_array.rb
@@ -15,8 +15,3 @@
 # prime_chars? ['a', 'bcd']       # => false
 # prime_chars? ['a', 'b', 'cd']   # => false
 
-require 'prime'
-
-def prime_chars?(strings)
-  Prime.prime?(strings.join.length)
-end
