+++ session2/3-challenge/7_array.rb
@@ -9,11 +9,3 @@
 # alternate_words("Can't we all get along?")      # => ["Can't", "all", "along"]
 # alternate_words("Elementary, my dear Watson!")  # => ["Elementary", "dear"]
 
-def alternate_words(strings)
-  strings = strings.gsub(/[^a-zA-Z_'’]/, " ")
-  myary = []
-  strings.split(" ").each_with_index do |word, index|
-    myary << word if index.even?
-  end
-  myary
-end
