+++ session2/3-challenge/9_input_output_logic_string.rb
@@ -23,20 +23,7 @@
 # GRANDMA:  HUH?! SPEAK UP, SONNY!
 # USER:     BYE
 
-def deaf_grandma
-  while line = gets
-    line.chomp!
-    if line == "BYE" && line != ""
-      break
-    elsif line == line.upcase && line != ""
-      puts "NO, NOT SINCE 1938!"
-      deaf_grandma
-    else
-      puts "HUH?! SPEAK UP, SONNY!"
-      deaf_grandma
-    end
-  end
-end
+
