+++ session2/5-solved/11.rb
@@ -2,34 +2,37 @@ class BeerSong
   attr_accessor :beers
 
   def initialize(beers)
-    beers = 0  if beers < 0
-    beers = 99 if beers > 99
-    self.beers = beers
+    if beers > 99
+      @beers = 99
+    elsif beers < 0
+      @beers = 0
+    else
+      @beers = beers
+    end
   end
 
   def print_song
-    beers.downto 1 do |i|
-      print_stanza i
+    @beers.downto 1 do |n|
+      print_verse n
     end
   end
 
-  def print_stanza(n)
-    if n.zero?
-      String.new
+  def print_verse(n)
+    if n == 0
+      return
     else
-      puts "#{translate n} #{bottle n} of beer on the wall,"        ,
-           "#{translate n} #{bottle n} of beer,"                    ,
-           "Take one down, pass it around,"                         ,
-           "#{translate n - 1} #{bottle n-1} of beer on the wall."
+      puts "#{translate(n)} #{bottle(n)} of beer on the wall,"
+      puts "#{translate(n)} #{bottle(n)} of beer,"
+      puts "Take one down, pass it around,"
+      puts "#{translate(n - 1)} #{bottle(n - 1)} of beer on the wall."
     end
   end
 
-  # returns "bottle" or "bottles"
   def bottle(n)
-    if n == 1 then 'bottle' else 'bottles' end
+    n == 1 ? "bottle" : "bottles"
   end
 
-  # translates number to English
+  # shamelessly stolen from the solution  
   def translate(n)
     if 0 <= n && n <= 19
       %w(zero one two three four five six seven eight nine ten eleven twelve thirteen fourteen fifteen sixteen seventeen eighteen nineteen)[n]
@@ -39,4 +42,5 @@ class BeerSong
       "#{translate n/10*10}-#{translate n%10}".downcase
     end.capitalize
   end
+    
 end
