+++ session2/5-solved/12.rb
@@ -1,25 +1,26 @@
 class Fraction
-  attr_accessor :numerator, :denominator
-
+  attr_accessor :numerator
+  attr_accessor :denominator
+  
   def initialize(numerator, denominator)
-    self.numerator, self.denominator = numerator, denominator
+    @numerator = numerator
+    @denominator = denominator
   end
-
+  
+  def to_f
+    @numerator.to_f / @denominator.to_f
+  end
+  
+  def lowest
+    Rational(@numerator, @denominator)
+  end
+  
   def to_s
     "#{numerator}/#{denominator}"
   end
-
-  def to_f
-    numerator / denominator.to_f
-  end
-
-  def gcd(a, b)
+  
+  def gcd(a,b)
     return a if b == 0
-    gcd b, (a % b)
-  end
-
-  def lowest
-    divisor = gcd(numerator, denominator)
-    Fraction.new(numerator/divisor, denominator/divisor)
+    gcd(b, a%b)
   end
 end
