+++ session2/5-solved/3.rb
@@ -1,9 +1,9 @@
 class String
   def every_other_char
-    to_return = ''
+    mystring = ""
     each_char.with_index do |char, index|
-      to_return << char if index.even?
+      index.even? ? mystring << char : nil
     end
-    to_return
+    mystring
   end
 end
