+++ session2/5-solved/7.rb
@@ -1,12 +1,8 @@
-def alternate_words(sentence)
-  # this will get better when we learn regular expressions :)
-  '!@$#%^&*()-=_+[]:;,./<>?\\|'.split(//).each do |char|
-    sentence = sentence.gsub(char, ' ')
+def alternate_words(strings)
+  strings = strings.gsub(/[^a-zA-Z_'’]/, " ")
+  myary = []
+  strings.split(" ").each_with_index do |word, index|
+    myary << word if index.even?
   end
-  words = sentence.split
-  solution = []
-  words.each_with_index do |word, index|
-    solution << word if index.even?
-  end
-  solution
+  myary
 end
