+++ session2/5-solved/9.rb
@@ -1,11 +1,14 @@
 def deaf_grandma
   while line = gets
     line.chomp!
-    break if line == "BYE"
-    if line == line.upcase && line != ""
+    if line == "BYE"
+      break
+    elsif line == line.upcase && line != ""
       puts "NO, NOT SINCE 1938!"
+      deaf_grandma
     else
       puts "HUH?! SPEAK UP, SONNY!"
+      deaf_grandma
     end
   end
 end
